refactor:코드 1차 분할
