+++ client/src/components/custom/RangeInputcontainer.tsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+const ContainerStyled = styled.div`
+  background-color: transparent;
+  position: absolute;
+  width: 100%;
+  bottom: 35%;
+  left: 0;
+  display: flex;
+  justify-content: center;
+  z-index: 15;
+`;
+
+export default ContainerStyled;
